:truck: Set the sentry plugin to be client only
